Added Restaurant

+++ backend/java/microservices/restaurant-service/src/main/java/com/pm/restaurantservice/service/RestaurantService.java
@@ -5,9 +5,6 @@ import com.cloudinary.utils.ObjectUtils;
 import com.pm.restaurantservice.dto.RestaurantRequestDTO;
 import com.pm.restaurantservice.dto.RestaurantResponseDTO;
 import com.pm.restaurantservice.exception.RestaurantNotFoundException;
-import com.pm.restaurantservice.grpc.AuthServiceGrpcClient;
-import com.pm.restaurantservice.grpc.BillingServiceGrpcClient;
-import com.pm.restaurantservice.kafka.KafkaProducer;
 import com.pm.restaurantservice.mapper.RestaurantMapper;
 import com.pm.restaurantservice.model.MenuItem;
 import com.pm.restaurantservice.model.Restaurant;
@@ -19,67 +16,69 @@ import java.io.IOException;
 import java.util.List;
 import java.util.Set;
 import java.util.UUID;
-
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
+import org.springframework.security.oauth2.core.oidc.OidcIdToken;
+import org.springframework.security.oauth2.core.oidc.user.DefaultOidcUser;
+import org.springframework.security.oauth2.jwt.Jwt;
+import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.server.ResponseStatusException;
 
 @Service
 public class RestaurantService {
-
   private final RestaurantRepository restaurantRepository;
-  private final BillingServiceGrpcClient billingServiceGrpcClient;
-  private final KafkaProducer kafkaProducer;
-  private final AuthServiceGrpcClient authServiceGrpcClient;
   private final Cloudinary cloudinary;
   public RestaurantService(RestaurantRepository restaurantRepository,
-      BillingServiceGrpcClient billingServiceGrpcClient,
-                           AuthServiceGrpcClient authServiceGrpcClient,
-      KafkaProducer kafkaProducer,Cloudinary cloudinary) {
+      Cloudinary cloudinary) {
     this.restaurantRepository = restaurantRepository;
-    this.billingServiceGrpcClient = billingServiceGrpcClient;
-    this.kafkaProducer = kafkaProducer;
-    this.authServiceGrpcClient=authServiceGrpcClient;
     this.cloudinary=cloudinary;
   }
 
   public List<RestaurantResponseDTO> getRestaurants() {
     List<Restaurant> restaurants = restaurantRepository.findAll();
-
     return restaurants.stream().map(RestaurantMapper::toDTO).toList();
   }
 
-  public String getAuth0Id(){
-    return authServiceGrpcClient.getAuth0Id();
-  }
-  public RestaurantResponseDTO createRestaurant(RestaurantRequestDTO restaurantRequestDTO,
-                                                @RequestParam("imageUrl") MultipartFile imageUrl) {
-    try {
-      File convFile = new File(System.getProperty("java.io.tmpdir") + "/" + imageUrl.getOriginalFilename());
-      FileOutputStream fos = new FileOutputStream(convFile);
-      fos.write(imageUrl.getBytes());
-      fos.close();
-
-      var pic = cloudinary.uploader().upload(convFile, ObjectUtils.asMap("folder", "/bookCovers/"));
-
-      Restaurant newRestaurant = new Restaurant();
-      newRestaurant.setImageUrl(pic.get("url").toString());
-              restaurantRepository.save(
-              RestaurantMapper.toModel(restaurantRequestDTO));
-
-      billingServiceGrpcClient.createBillingAccount(newRestaurant.getId().toString(),
-              newRestaurant.getRestaurantName(), newRestaurant.getRestaurantName());
-
-      kafkaProducer.sendEvent(newRestaurant);
-
-      return RestaurantMapper.toDTO(newRestaurant);
-
-    } catch (IOException e) {
-      throw new ResponseStatusException(HttpStatus.BAD_GATEWAY, "Failed to upload the file.");
+  public static String getAuth0Id(Authentication authentication) {
+     if (authentication instanceof OAuth2AuthenticationToken token && token.getPrincipal() instanceof DefaultOidcUser user) {
+      OidcIdToken idToken = user.getIdToken();
+      return idToken.getClaimAsString("sub");
     }
+    else if (authentication instanceof JwtAuthenticationToken jwtAuthenticationToken) {
+      Jwt jwt = jwtAuthenticationToken.getToken();
+      return jwt.getClaimAsString("sub");
+    }
+    else{
+      return null;
+    }
+  }
+  public RestaurantResponseDTO createRestaurant(@RequestPart RestaurantRequestDTO restaurantRequestDTO,
+                                                @RequestPart(value = "imageFile") MultipartFile imageFile,
+                                                Authentication authentication) {
+    if(restaurantRepository.existsByUserId(getAuth0Id(authentication))){
+      return RestaurantMapper.toDTO(null);
+    }
+    else{
+      try {
+        File convFile = new File(System.getProperty("java.io.tmpdir") + "/" + imageFile.getOriginalFilename());
+        FileOutputStream fos = new FileOutputStream(convFile);
+        fos.write(imageFile.getBytes());
+        fos.close();
 
+        var pic = cloudinary.uploader().upload(convFile, ObjectUtils.asMap("folder", "/mern-food-ordering-app/"));
+        restaurantRequestDTO.setImageUrl(pic.get("url").toString());
+        Restaurant newRestaurant=restaurantRepository.save(
+                RestaurantMapper.toModel(restaurantRequestDTO));
+        return RestaurantMapper.toDTO(newRestaurant);
+
+      } catch (IOException e) {
+        throw new ResponseStatusException(HttpStatus.BAD_GATEWAY, "Failed to upload the file.");
+      }
+    }
   }
 
   public RestaurantResponseDTO updateRestaurant(UUID id,
